refactor: split auth cookie and provider internals
