+++ src/stdlib/auth.rs
@@ -43,6 +43,26 @@ use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 use totp_rs::{Algorithm as TotpAlgorithm, Secret, TOTP};
 
+mod cookies;
+mod providers;
+
+pub use cookies::default_auth_cookie_secure_env;
+use cookies::{
+    add_set_cookie_header, auth_challenge_cookie_name, build_cleared_auth_challenge_cookie,
+    build_cleared_session_cookie, build_signed_auth_challenge_cookie, build_signed_session_cookie,
+};
+use providers::{
+    available_providers, get_builtin_provider, is_safe_provider_name, suggest_provider,
+    validate_provider_name, value_map_to_provider,
+};
+pub use providers::{provider_to_value, value_to_provider};
+
+// Internal architecture note:
+// Phase 4.5 is being executed in small, behavior-preserving slices. `auth.rs` remains the
+// public std/auth surface for now, while lower-risk internals move into focused submodules.
+// During this cleanup, prefer explicit semantics over clever abstraction, keep fallback/error
+// behavior intentional, and avoid mixing new auth features into refactor-only slices.
+
 // ============================================================================
 // SECURITY HELPERS
 // ============================================================================
@@ -260,172 +280,6 @@ pub struct OidcDiscovery {
     pub scopes_supported: Vec<String>,
     pub response_types_supported: Vec<String>,
     pub token_endpoint_auth_methods_supported: Vec<String>,
-}
-
-// ============================================================================
-// SECTION 3: Built-in Provider Registry
-// ============================================================================
-
-/// Built-in provider template (without credentials)
-struct BuiltinProvider {
-    name: &'static str,
-    authorize_url: &'static str,
-    token_url: &'static str,
-    userinfo_url: &'static str,
-    issuer: Option<&'static str>,
-    default_scopes: &'static [&'static str],
-    supports_oidc: bool,
-    supports_pkce: bool,
-}
-
-static BUILTIN_PROVIDERS: &[BuiltinProvider] = &[
-    BuiltinProvider {
-        name: "google",
-        authorize_url: "https://accounts.google.com/o/oauth2/v2/auth",
-        token_url: "https://oauth2.googleapis.com/token",
-        userinfo_url: "https://www.googleapis.com/oauth2/v2/userinfo",
-        issuer: Some("https://accounts.google.com"),
-        default_scopes: &["openid", "email", "profile"],
-        supports_oidc: true,
-        supports_pkce: true,
-    },
-    BuiltinProvider {
-        name: "github",
-        authorize_url: "https://github.com/login/oauth/authorize",
-        token_url: "https://github.com/login/oauth/access_token",
-        userinfo_url: "https://api.github.com/user",
-        issuer: None,
-        default_scopes: &["read:user", "user:email"],
-        supports_oidc: false,
-        supports_pkce: false,
-    },
-    BuiltinProvider {
-        name: "facebook",
-        authorize_url: "https://www.facebook.com/v18.0/dialog/oauth",
-        token_url: "https://graph.facebook.com/v18.0/oauth/access_token",
-        userinfo_url: "https://graph.facebook.com/me?fields=id,name,email,picture",
-        issuer: None,
-        default_scopes: &["email", "public_profile"],
-        supports_oidc: false,
-        supports_pkce: true,
-    },
-    BuiltinProvider {
-        name: "microsoft",
-        authorize_url: "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
-        token_url: "https://login.microsoftonline.com/common/oauth2/v2.0/token",
-        userinfo_url: "https://graph.microsoft.com/v1.0/me",
-        issuer: Some("https://login.microsoftonline.com/common/v2.0"),
-        default_scopes: &["openid", "email", "profile"],
-        supports_oidc: true,
-        supports_pkce: true,
-    },
-    BuiltinProvider {
-        name: "discord",
-        authorize_url: "https://discord.com/api/oauth2/authorize",
-        token_url: "https://discord.com/api/oauth2/token",
-        userinfo_url: "https://discord.com/api/users/@me",
-        issuer: None,
-        default_scopes: &["identify", "email"],
-        supports_oidc: false,
-        supports_pkce: true,
-    },
-    BuiltinProvider {
-        name: "twitter",
-        authorize_url: "https://twitter.com/i/oauth2/authorize",
-        token_url: "https://api.twitter.com/2/oauth2/token",
-        userinfo_url: "https://api.twitter.com/2/users/me?user.fields=profile_image_url",
-        issuer: None,
-        default_scopes: &["users.read", "tweet.read"],
-        supports_oidc: false,
-        supports_pkce: true, // Twitter OAuth 2.0 requires PKCE
-    },
-    BuiltinProvider {
-        name: "linkedin",
-        authorize_url: "https://www.linkedin.com/oauth/v2/authorization",
-        token_url: "https://www.linkedin.com/oauth/v2/accessToken",
-        userinfo_url: "https://api.linkedin.com/v2/userinfo",
-        issuer: Some("https://www.linkedin.com"),
-        default_scopes: &["openid", "profile", "email"],
-        supports_oidc: true,
-        supports_pkce: false,
-    },
-    BuiltinProvider {
-        name: "apple",
-        authorize_url: "https://appleid.apple.com/auth/authorize",
-        token_url: "https://appleid.apple.com/auth/token",
-        userinfo_url: "", // Apple uses ID token, no userinfo endpoint
-        issuer: Some("https://appleid.apple.com"),
-        default_scopes: &["name", "email"],
-        supports_oidc: true,
-        supports_pkce: false,
-    },
-    BuiltinProvider {
-        name: "okta",
-        authorize_url: "", // Requires tenant configuration
-        token_url: "",
-        userinfo_url: "",
-        issuer: None, // Dynamic based on tenant
-        default_scopes: &["openid", "email", "profile"],
-        supports_oidc: true,
-        supports_pkce: true,
-    },
-    BuiltinProvider {
-        name: "auth0",
-        authorize_url: "", // Requires domain configuration
-        token_url: "",
-        userinfo_url: "",
-        issuer: None, // Dynamic based on domain
-        default_scopes: &["openid", "email", "profile"],
-        supports_oidc: true,
-        supports_pkce: true,
-    },
-];
-
-/// Get a built-in provider by name
-fn get_builtin_provider(name: &str) -> Option<&'static BuiltinProvider> {
-    BUILTIN_PROVIDERS.iter().find(|p| p.name == name)
-}
-
-/// Get list of available provider names (for error messages)
-fn available_providers() -> String {
-    BUILTIN_PROVIDERS
-        .iter()
-        .filter(|p| !p.authorize_url.is_empty()) // Only fully configured providers
-        .map(|p| p.name)
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-/// Find similar provider name for typo suggestions
-fn suggest_provider(name: &str) -> Option<&'static str> {
-    BUILTIN_PROVIDERS
-        .iter()
-        .map(|p| p.name)
-        .min_by_key(|p| levenshtein_distance(name, p))
-        .filter(|p| levenshtein_distance(name, p) <= 2)
-}
-
-fn levenshtein_distance(a: &str, b: &str) -> usize {
-    let a: Vec<char> = a.chars().collect();
-    let b: Vec<char> = b.chars().collect();
-    let mut dp = vec![vec![0; b.len() + 1]; a.len() + 1];
-
-    for (i, item) in dp.iter_mut().enumerate().take(a.len() + 1) {
-        item[0] = i;
-    }
-    for j in 0..=b.len() {
-        dp[0][j] = j;
-    }
-
-    for i in 1..=a.len() {
-        for j in 1..=b.len() {
-            let cost = if a[i - 1] == b[j - 1] { 0 } else { 1 };
-            dp[i][j] = (dp[i - 1][j] + 1)
-                .min(dp[i][j - 1] + 1)
-                .min(dp[i - 1][j - 1] + cost);
-        }
-    }
-    dp[a.len()][b.len()]
 }
 
 /// Extract user info from provider response or ID token
@@ -1455,26 +1309,6 @@ fn initialize_session_store(store: &SessionStore) -> std::result::Result<(), Str
     }
 }
 
-pub fn default_auth_cookie_secure_env() -> bool {
-    if let Ok(env) = std::env::var("NTNT_ENV") {
-        if env.eq_ignore_ascii_case("development") || env.eq_ignore_ascii_case("dev") {
-            return false;
-        }
-    }
-
-    if let Ok(site_url) = std::env::var("SITE_URL") {
-        let lower = site_url.to_ascii_lowercase();
-        if lower.starts_with("http://localhost")
-            || lower.starts_with("http://127.0.0.1")
-            || lower.starts_with("http://0.0.0.0")
-        {
-            return false;
-        }
-    }
-
-    true
-}
-
 pub fn parse_auth_session_store(value: &str) -> std::result::Result<SessionStore, String> {
     if value == "memory" || value.is_empty() {
         return Ok(SessionStore::Memory);
@@ -1657,24 +1491,6 @@ fn with_protected_paths<T>(f: impl FnOnce(&[String]) -> T) -> T {
     f(&protected)
 }
 
-fn is_safe_provider_name(name: &str) -> bool {
-    !name.is_empty()
-        && name
-            .chars()
-            .all(|ch| ch.is_ascii_alphanumeric() || ch == '_' || ch == '-' || ch == '.')
-}
-
-fn validate_provider_name(name: &str) -> std::result::Result<(), String> {
-    if is_safe_provider_name(name) {
-        Ok(())
-    } else {
-        Err(
-            "provider name must use only ASCII letters, numbers, periods, underscores, or hyphens"
-                .to_string(),
-        )
-    }
-}
-
 fn validate_auth_challenge_kind(kind: &str) -> std::result::Result<String, String> {
     if kind.is_empty() {
         return Err("[auth] begin_auth_challenge() kind must not be empty".to_string());
@@ -1707,255 +1523,6 @@ fn encode_url_path_segment(text: &str) -> String {
     }
 
     encoded
-}
-
-#[derive(Debug, Clone)]
-struct AuthCookieSettings {
-    name: String,
-    path: String,
-    same_site: String,
-    secure: bool,
-    http_only: bool,
-    max_age: i64,
-}
-
-fn default_auth_cookie_max_age(config: &AuthConfig) -> i64 {
-    if config.store_tokens && config.refresh_ttl > config.session_ttl {
-        config.refresh_ttl
-    } else {
-        config.session_ttl
-    }
-}
-
-fn validate_cookie_name(name: &str) -> std::result::Result<String, String> {
-    if name.is_empty() {
-        return Err("[auth] cookie_name must not be empty".to_string());
-    }
-
-    if name.chars().all(|ch| {
-        ch.is_ascii_alphanumeric()
-            || matches!(
-                ch,
-                '!' | '#'
-                    | '$'
-                    | '%'
-                    | '&'
-                    | '\''
-                    | '*'
-                    | '+'
-                    | '-'
-                    | '.'
-                    | '^'
-                    | '_'
-                    | '`'
-                    | '|'
-                    | '~'
-            )
-    }) {
-        Ok(name.to_string())
-    } else {
-        Err("[auth] cookie_name contains invalid characters".to_string())
-    }
-}
-
-fn validate_cookie_path(path: &str) -> std::result::Result<String, String> {
-    if path.is_empty() {
-        return Err("[auth] cookie_path must not be empty".to_string());
-    }
-    if !path.starts_with('/') {
-        return Err("[auth] cookie_path must start with /".to_string());
-    }
-    if path.chars().any(|ch| ch.is_control() || ch == ';') {
-        return Err("[auth] cookie_path contains invalid characters".to_string());
-    }
-
-    Ok(path.to_string())
-}
-
-fn normalize_cookie_same_site(value: &str) -> std::result::Result<String, String> {
-    match value.to_ascii_lowercase().as_str() {
-        "lax" => Ok("Lax".to_string()),
-        "strict" => Ok("Strict".to_string()),
-        "none" => Ok("None".to_string()),
-        _ => Err("[auth] cookie_same_site must be one of: Lax, Strict, None".to_string()),
-    }
-}
-
-fn auth_cookie_settings(
-    config: &AuthConfig,
-    max_age: i64,
-    overrides: Option<&HashMap<String, Value>>,
-) -> std::result::Result<AuthCookieSettings, String> {
-    if overrides.and_then(|m| m.get("cookie_name")).is_some() {
-        return Err(
-            "[auth] cookie_name override is not supported; configure it via enable_auth()"
-                .to_string(),
-        );
-    }
-
-    let get_string = |key: &str| -> std::result::Result<Option<String>, String> {
-        match overrides.and_then(|m| m.get(key)) {
-            Some(Value::String(s)) => Ok(Some(s.clone())),
-            Some(other) => Err(format!(
-                "[auth] {} must be a string, got {}",
-                key,
-                other.type_name()
-            )),
-            None => Ok(None),
-        }
-    };
-
-    let get_bool = |key: &str| -> std::result::Result<Option<bool>, String> {
-        match overrides.and_then(|m| m.get(key)) {
-            Some(Value::Bool(b)) => Ok(Some(*b)),
-            Some(other) => Err(format!(
-                "[auth] {} must be a bool, got {}",
-                key,
-                other.type_name()
-            )),
-            None => Ok(None),
-        }
-    };
-
-    let cookie_max_age = match overrides.and_then(|m| m.get("cookie_max_age")) {
-        Some(Value::Int(i)) => *i,
-        Some(other) => {
-            return Err(format!(
-                "[auth] cookie_max_age must be an int, got {}",
-                other.type_name()
-            ))
-        }
-        None => max_age,
-    };
-
-    Ok(AuthCookieSettings {
-        name: validate_cookie_name(&config.cookie_name)?,
-        path: validate_cookie_path(&get_string("cookie_path")?.unwrap_or_else(|| "/".to_string()))?,
-        same_site: normalize_cookie_same_site(
-            &get_string("cookie_same_site")?.unwrap_or_else(|| config.cookie_same_site.clone()),
-        )?,
-        secure: get_bool("cookie_secure")?.unwrap_or(config.cookie_secure),
-        http_only: get_bool("cookie_http_only")?.unwrap_or(true),
-        max_age: cookie_max_age,
-    })
-}
-
-fn build_auth_cookie_string(value: &str, settings: &AuthCookieSettings) -> String {
-    let mut cookie = format!(
-        "{}={}; Path={}; Max-Age={}; SameSite={}",
-        settings.name, value, settings.path, settings.max_age, settings.same_site
-    );
-
-    if settings.http_only {
-        cookie.push_str("; HttpOnly");
-    }
-    if settings.secure {
-        cookie.push_str("; Secure");
-    }
-
-    cookie
-}
-
-fn build_signed_session_cookie(
-    config: &AuthConfig,
-    session_id: &str,
-    overrides: Option<&HashMap<String, Value>>,
-) -> std::result::Result<String, String> {
-    let settings = auth_cookie_settings(config, default_auth_cookie_max_age(config), overrides)?;
-    let signed_session_id = sign_session_id(session_id, &config.session_secret);
-    Ok(build_auth_cookie_string(&signed_session_id, &settings))
-}
-
-fn build_cleared_session_cookie(
-    config: &AuthConfig,
-    overrides: Option<&HashMap<String, Value>>,
-) -> std::result::Result<String, String> {
-    let sanitized_overrides = overrides.map(|map| {
-        let mut sanitized = map.clone();
-        sanitized.remove("cookie_max_age");
-        sanitized
-    });
-    let settings = auth_cookie_settings(config, 0, sanitized_overrides.as_ref())?;
-    Ok(build_auth_cookie_string("", &settings))
-}
-
-fn auth_challenge_cookie_name(config: &AuthConfig) -> std::result::Result<String, String> {
-    validate_cookie_name(&format!("{}_challenge", config.cookie_name))
-}
-
-fn build_signed_auth_challenge_cookie(
-    config: &AuthConfig,
-    challenge_id: &str,
-    ttl: i64,
-) -> std::result::Result<String, String> {
-    let mut settings = auth_cookie_settings(config, ttl.max(0), None)?;
-    settings.name = auth_challenge_cookie_name(config)?;
-    let signed_challenge_id = sign_session_id(challenge_id, &config.session_secret);
-    Ok(build_auth_cookie_string(&signed_challenge_id, &settings))
-}
-
-fn build_cleared_auth_challenge_cookie(config: &AuthConfig) -> std::result::Result<String, String> {
-    let mut settings = auth_cookie_settings(config, 0, None)?;
-    settings.name = auth_challenge_cookie_name(config)?;
-    Ok(build_auth_cookie_string("", &settings))
-}
-
-fn add_set_cookie_header(response: &Value, cookie: &str) -> std::result::Result<Value, String> {
-    let mut response = match response {
-        Value::Map(map) => map.clone(),
-        other => {
-            return Err(format!(
-                "[auth] response must be a map, got {}",
-                other.type_name()
-            ))
-        }
-    };
-
-    let mut headers = match response.get("headers") {
-        Some(Value::Map(headers)) => headers.clone(),
-        Some(other) => {
-            return Err(format!(
-                "[auth] response headers must be a map, got {}",
-                other.type_name()
-            ))
-        }
-        None => HashMap::new(),
-    };
-
-    let set_cookie_key = headers
-        .keys()
-        .find(|key| key.eq_ignore_ascii_case("set-cookie"))
-        .cloned()
-        .unwrap_or_else(|| "Set-Cookie".to_string());
-
-    match headers.get(&set_cookie_key) {
-        Some(Value::Array(existing)) => {
-            let mut cookies = existing.clone();
-            cookies.push(Value::String(cookie.to_string()));
-            headers.insert(set_cookie_key, Value::Array(cookies));
-        }
-        Some(Value::String(existing)) => {
-            headers.insert(
-                set_cookie_key,
-                Value::Array(vec![
-                    Value::String(existing.clone()),
-                    Value::String(cookie.to_string()),
-                ]),
-            );
-        }
-        Some(other) => {
-            return Err(format!(
-                "[auth] response Set-Cookie header must be a string or array, got {}",
-                other.type_name()
-            ))
-        }
-        None => {
-            headers.insert(set_cookie_key, Value::String(cookie.to_string()));
-        }
-    }
-
-    response.insert("headers".to_string(), Value::Map(headers));
-    Ok(Value::Map(response))
 }
 
 pub fn reset_protected_paths() {
@@ -5618,77 +5185,6 @@ fn json_response(data: Value, status: i64) -> Value {
     Value::Map(response)
 }
 
-/// Convert a Value::Map (from oauth()) to a ProviderConfig
-fn value_map_to_provider(
-    map: &HashMap<String, Value>,
-) -> std::result::Result<ProviderConfig, String> {
-    let get_str = |key: &str| -> Option<String> {
-        map.get(key).and_then(|v| match v {
-            Value::String(s) => Some(s.clone()),
-            _ => None,
-        })
-    };
-
-    let get_bool = |key: &str| -> bool {
-        map.get(key)
-            .and_then(|v| match v {
-                Value::Bool(b) => Some(*b),
-                _ => None,
-            })
-            .unwrap_or(false)
-    };
-
-    let get_str_array = |key: &str| -> Vec<String> {
-        map.get(key)
-            .and_then(|v| match v {
-                Value::Array(arr) => Some(
-                    arr.iter()
-                        .filter_map(|v| match v {
-                            Value::String(s) => Some(s.clone()),
-                            _ => None,
-                        })
-                        .collect(),
-                ),
-                _ => None,
-            })
-            .unwrap_or_default()
-    };
-
-    let name = get_str("name").ok_or("Provider must have a 'name'")?;
-    let authorize_url = get_str("authorize_url").ok_or("Provider must have 'authorize_url'")?;
-    let token_url = get_str("token_url").ok_or("Provider must have 'token_url'")?;
-    let client_id = get_str("client_id").ok_or("Provider must have 'client_id'")?;
-
-    validate_provider_name(&name)?;
-
-    Ok(ProviderConfig {
-        name,
-        client_id,
-        client_secret: get_str("client_secret").unwrap_or_default(),
-        authorize_url,
-        token_url,
-        userinfo_url: get_str("userinfo_url").unwrap_or_default(),
-        scopes: get_str_array("scopes"),
-        supports_oidc: get_bool("supports_oidc"),
-        issuer: get_str("issuer"),
-        jwks_uri: get_str("jwks_uri"),
-        use_pkce: get_bool("use_pkce"),
-        extra_params: {
-            let mut params = HashMap::new();
-            // Parse extra_params from config if provided
-            if let Some(Value::Map(ep)) = map.get("extra_params") {
-                for (k, v) in ep {
-                    if let Value::String(s) = v {
-                        params.insert(k.clone(), s.clone());
-                    }
-                }
-            }
-            params
-        },
-    })
-}
-
-/// Handle GET /auth/{provider} - Start OAuth flow
 pub fn handle_auth_start(args: &[Value]) -> Result<Value> {
     let req = &args[0];
 
@@ -6125,155 +5621,6 @@ pub fn handle_auth_logout(args: &[Value]) -> Result<Value> {
 // ============================================================================
 
 /// Convert ProviderConfig to Value::Map
-pub fn provider_to_value(provider: &ProviderConfig) -> Value {
-    let mut map = HashMap::new();
-    map.insert("_provider".to_string(), Value::Bool(true));
-    map.insert("name".to_string(), Value::String(provider.name.clone()));
-    map.insert(
-        "client_id".to_string(),
-        Value::String(provider.client_id.clone()),
-    );
-    map.insert(
-        "client_secret".to_string(),
-        Value::String(provider.client_secret.clone()),
-    );
-    map.insert(
-        "authorize_url".to_string(),
-        Value::String(provider.authorize_url.clone()),
-    );
-    map.insert(
-        "token_url".to_string(),
-        Value::String(provider.token_url.clone()),
-    );
-    map.insert(
-        "userinfo_url".to_string(),
-        Value::String(provider.userinfo_url.clone()),
-    );
-    map.insert(
-        "scopes".to_string(),
-        Value::Array(
-            provider
-                .scopes
-                .iter()
-                .map(|s| Value::String(s.clone()))
-                .collect(),
-        ),
-    );
-    map.insert("use_pkce".to_string(), Value::Bool(provider.use_pkce));
-    map.insert(
-        "supports_oidc".to_string(),
-        Value::Bool(provider.supports_oidc),
-    );
-
-    if let Some(ref issuer) = provider.issuer {
-        map.insert("issuer".to_string(), Value::String(issuer.clone()));
-    }
-    if !provider.extra_params.is_empty() {
-        let mut extra = HashMap::new();
-        for (k, v) in &provider.extra_params {
-            extra.insert(k.clone(), Value::String(v.clone()));
-        }
-        map.insert("extra_params".to_string(), Value::Map(extra));
-    }
-    Value::Map(map)
-}
-
-/// Parse ProviderConfig from Value::Map
-pub fn value_to_provider(value: &Value) -> Result<ProviderConfig> {
-    match value {
-        Value::Map(map) => {
-            let get_str = |key: &str| -> Result<String> {
-                match map.get(key) {
-                    Some(Value::String(s)) => Ok(s.clone()),
-                    _ => Err(IntentError::type_error(format!(
-                        "Provider {} must be a string",
-                        key
-                    ))),
-                }
-            };
-
-            let get_bool = |key: &str, default: bool| -> bool {
-                match map.get(key) {
-                    Some(Value::Bool(b)) => *b,
-                    _ => default,
-                }
-            };
-
-            let name = get_str("name")?;
-            validate_provider_name(&name).map_err(IntentError::type_error)?;
-            let client_id = get_str("client_id")?;
-            let client_secret = get_str("client_secret")?;
-            let authorize_url = get_str("authorize_url")?;
-            let token_url = get_str("token_url")?;
-            let userinfo_url = map
-                .get("userinfo_url")
-                .and_then(|v| {
-                    if let Value::String(s) = v {
-                        Some(s.clone())
-                    } else {
-                        None
-                    }
-                })
-                .unwrap_or_default();
-
-            let issuer = map.get("issuer").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-
-            let jwks_uri = map.get("jwks_uri").and_then(|v| {
-                if let Value::String(s) = v {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-
-            let scopes = match map.get("scopes") {
-                Some(Value::Array(arr)) => arr
-                    .iter()
-                    .filter_map(|v| match v {
-                        Value::String(s) => Some(s.clone()),
-                        _ => None,
-                    })
-                    .collect(),
-                _ => vec![],
-            };
-
-            let extra_params = match map.get("extra_params") {
-                Some(Value::Map(m)) => m
-                    .iter()
-                    .filter_map(|(k, v)| match v {
-                        Value::String(s) => Some((k.clone(), s.clone())),
-                        _ => None,
-                    })
-                    .collect(),
-                _ => HashMap::new(),
-            };
-
-            Ok(ProviderConfig {
-                name,
-                client_id,
-                client_secret,
-                authorize_url,
-                token_url,
-                userinfo_url,
-                issuer,
-                jwks_uri,
-                scopes,
-                extra_params,
-                use_pkce: get_bool("use_pkce", false),
-                supports_oidc: get_bool("supports_oidc", false),
-            })
-        }
-        _ => Err(IntentError::type_error(
-            "Provider must be a map".to_string(),
-        )),
-    }
-}
 
 // ============================================================================
 // SECTION 11: Module Initialization
